requirements.txt
absl
numpy
matplotlib
pandas
keras
tensorflow
scikit-learn
Pillow
opencv-python
scikit-image
openslide-python
coloredlogs
